# docs/api/openapi-selfhosted-byom-v1.yaml
openapi: 3.0.3
info:
  title: Log RCA Copilot Self-hosted API
  version: 1.0.0-mvp
  description: |
    MVP API contract for self-hosted BYOM architecture.
    Scope: Loki integration, incident analysis, OpenAI/Gemini OAuth+API key,
    Slack/Email alerts, and operational endpoints.
servers:
  - url: http://localhost:8080
    description: Local self-hosted server
...
paths:
...
  /v1/projects/{project_id}/policies/export:
    get:
      tags: [Policies]
      summary: Get export policy
      operationId: getExportPolicy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Current export policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPolicyDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
    put:
      tags: [Policies]
      summary: Update export policy
      operationId: updateExportPolicy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportPolicyRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportPolicyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/projects/{project_id}/policies/redaction:
    get:
      tags: [Policies]
      summary: Get redaction policy
      operationId: getRedactionPolicy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Current redaction policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedactionPolicyDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
    put:
      tags: [Policies]
      summary: Update masking/redaction rules
      operationId: updateRedactionPolicy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedactionPolicyRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedactionPolicyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/projects/{project_id}/alerts/slack:
    get:
      tags: [Alerts]
      summary: Get Slack alert channel configuration
      operationId: getSlackConfiguration
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Slack configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlackAlertConfigResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Alerts]
      summary: Configure Slack alert channel
      operationId: configureSlack
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlackAlertRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertChannelResponse'
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertChannelResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /v1/projects/{project_id}/alerts/email:
    get:
      tags: [Alerts]
      summary: Get Email alert channel configuration
      operationId: getEmailConfiguration
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Email configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailAlertConfigResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Alerts]
      summary: Configure Email alert channel
      operationId: configureEmail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailAlertRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertChannelResponse'
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertChannelResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /v1/projects/{project_id}/audit-logs:
    get:
      tags: [Audit]
      summary: List project audit logs
      operationId: listAuditLogs
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: action
          in: query
          schema:
            type: string
        - name: actor
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ingestToken:
      type: http
      scheme: bearer
      bearerFormat: IngestToken
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ApiMeta:
      type: object
      properties:
        request_id:
          type: string
        next_cursor:
          type: string
          nullable: true
      additionalProperties: false
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: true
          required: [code, message]
      required: [error]
...
    ExportPolicyRequest:
      type: object
      properties:
        level:
          type: string
          enum: [level0_rule_only, level1_byom_only, level2_byom_with_telemetry]
      required: [level]
    ExportPolicyResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            level:
              type: string
            updated_at:
              type: string
              format: date-time
          required: [level, updated_at]
      required: [data]
    ExportPolicyDetailResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            configured:
              type: boolean
            level:
              type: string
              enum: [level0_rule_only, level1_byom_only, level2_byom_with_telemetry]
            updated_at:
              type: string
              format: date-time
              nullable: true
          required: [configured, level]
      required: [data]
    RedactionRule:
      type: object
      properties:
        name:
          type: string
        pattern:
          type: string
        replace_with:
          type: string
      required: [name, pattern, replace_with]
    RedactionPolicyRequest:
      type: object
      properties:
        enabled:
          type: boolean
        rules:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/RedactionRule'
      required: [enabled, rules]
    RedactionPolicyResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            enabled:
              type: boolean
            rules_count:
              type: integer
            updated_at:
              type: string
              format: date-time
          required: [enabled, rules_count, updated_at]
      required: [data]
    RedactionPolicyDetailResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            configured:
              type: boolean
            enabled:
              type: boolean
            rules_count:
              type: integer
            rules:
              type: array
              items:
                $ref: '#/components/schemas/RedactionRule'
            updated_at:
              type: string
              format: date-time
              nullable: true
          required: [configured, enabled, rules_count, rules]
      required: [data]
...
    SlackAlertRequest:
      type: object
      properties:
        webhook_url:
          oneOf:
            - type: string
              format: uri
            - type: string
              maxLength: 0
              description: Empty string keeps previously stored webhook secret when already configured.
          nullable: true
        channel:
          type: string
        min_confidence:
          type: number
          minimum: 0
          maximum: 1
          default: 0.45
      required: [channel]
    EmailAlertRequest:
      type: object
      properties:
        from:
          type: string
          format: email
        recipients:
          type: array
          minItems: 1
          items:
            type: string
            format: email
        smtp:
          type: object
          properties:
            host:
              type: string
            port:
              type: integer
            username:
              type: string
            password:
              oneOf:
                - type: string
                - type: string
                  maxLength: 0
                  description: Empty string keeps previously stored SMTP password when already configured.
              nullable: true
            starttls:
              type: boolean
              default: true
          required: [host, port, username]
        min_confidence:
          type: number
          minimum: 0
          maximum: 1
          default: 0.45
      required: [from, recipients, smtp]
    AlertChannelResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
            type:
              type: string
              enum: [slack, email]
            enabled:
              type: boolean
            updated_at:
              type: string
              format: date-time
          required: [id, type, enabled, updated_at]
      required: [data]
    SlackAlertConfigResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            configured:
              type: boolean
            id:
              type: string
              nullable: true
            type:
              type: string
              enum: [slack]
            enabled:
              type: boolean
            webhook_url:
              type: string
              format: uri
              nullable: true
            channel:
              type: string
              nullable: true
            webhook_configured:
              type: boolean
            min_confidence:
              type: number
              minimum: 0
              maximum: 1
            updated_at:
              type: string
              format: date-time
              nullable: true
          required: [configured, type, enabled, webhook_configured, min_confidence]
      required: [data]
    EmailAlertConfigResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            configured:
              type: boolean
            id:
              type: string
              nullable: true
            type:
              type: string
              enum: [email]
            enabled:
              type: boolean
            from:
              type: string
              format: email
              nullable: true
            recipients:
              type: array
              items:
                type: string
                format: email
            smtp:
              $ref: '#/components/schemas/SmtpConfigData'
            min_confidence:
              type: number
              minimum: 0
              maximum: 1
            updated_at:
              type: string
              format: date-time
              nullable: true
          required: [configured, type, enabled, recipients, smtp, min_confidence]
      required: [data]
    SmtpConfigData:
      type: object
      properties:
        host:
          type: string
          nullable: true
        port:
          type: integer
          nullable: true
        username:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        password_configured:
          type: boolean
        starttls:
          type: boolean
      required: [password_configured, starttls]
    AuditLog:
      type: object
      properties:
        id:
          type: string
        actor:
          type: string
        action:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        created_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
      required: [id, actor, action, resource_type, resource_id, created_at]
    AuditLogListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        meta:
          $ref: '#/components/schemas/ApiMeta'
      required: [data]